flowedit: Extract flow_io module from main.rs
Move flow file operations into a dedicated flow_io.rs module:
- perform_save/save_as/open/new/compile
- load_flow, save_flow_toml, and TOML serialization helpers
- build_meta_provider, resolve_lib_paths, extract_ports
- save_function_definition
- editor preferences (save/load/path)
- generate_unique_alias, next_node_position, format_endpoint

Associated unit tests moved to flow_io::test module.
